Replace parser code
