+++ openapi_python_client/cli_endpoint_selection.py
@@ -2,8 +2,7 @@ from typing import List, Set
 
 import questionary
 
-# from .parser.endpoint_collection import Endpoints, Endpoint
-from parser.endpoints import Endpoint, EndpointCollection
+from openapi_python_client.parser.endpoints import Endpoint, EndpointCollection
 
 
 def questionary_endpoint_selection(endpoints: EndpointCollection) -> Set[str]:
@@ -14,12 +13,6 @@ def questionary_endpoint_selection(endpoints: EndpointCollection) -> Set[str]:
         if prev_table_name != endpoint.table_name:
             choices.append(questionary.Separator(f"\n{endpoint.table_name} endpoints:\n"))
         prev_table_name = endpoint.table_name
-        # for tag, collection in endpoints.endpoints_by_tag.items():
-        #     if not collection.endpoints_to_render:
-        #         continue
-        #     choices.append(questionary.Separator(f"\n{tag} endpoints:\n"))
-        # for endpoint in collection.endpoints_to_render:
-        # text = [("bold", str(endpoint.python_name))]  # , ("italic fg:ansigray", f" {endpoint.path}")]
         text = [
             ("bold", str(endpoint.python_name)),
             ("italic", f" {endpoint.path}"),
