--- openapi_python_client/parser/__init__.py
@@ -1,6 +1,0 @@
-""" Classes representing the data in the OpenAPI schema """
-
-__all__ = ["GeneratorData", "import_string_from_class"]
-
-from .endpoints import import_string_from_class
-from .openapi import GeneratorData
