+++ openapi_python_client/parser/paths.py
