--- openapi_python_client/parser/properties/types.py
@@ -1,143 +1,0 @@
-from typing import ClassVar, Optional, Set
-
-import attr
-
-from ... import schema as oai
-from .list_property import ListProperty
-from .none_property import NoneProperty
-from .property import Property
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class AnyProperty(Property):
-    """A property that can be any type (used for empty schemas)"""
-
-    _type_string: ClassVar[str] = "Any"
-    _json_type_string: ClassVar[str] = "Any"
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class StringProperty(Property):
-    """A property of type str"""
-
-    max_length: Optional[int] = None
-    pattern: Optional[str] = None
-    _type_string: ClassVar[str] = "str"
-    _json_type_string: ClassVar[str] = "str"
-    _allowed_locations: ClassVar[Set[oai.ParameterLocation]] = {
-        oai.ParameterLocation.QUERY,
-        oai.ParameterLocation.PATH,
-        oai.ParameterLocation.COOKIE,
-        oai.ParameterLocation.HEADER,
-    }
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class DateTimeProperty(Property):
-    """
-    A property of type datetime.datetime
-    """
-
-    _type_string: ClassVar[str] = "datetime.datetime"
-    _json_type_string: ClassVar[str] = "str"
-    template: ClassVar[str] = "datetime_property.py.jinja"
-
-    def get_imports(self, *, prefix: str) -> Set[str]:
-        """
-        Get a set of import strings that should be included when this property is used somewhere
-
-        Args:
-            prefix: A prefix to put before any relative (local) module names. This should be the number of . to get
-            back to the root of the generated client.
-        """
-        imports = super().get_imports(prefix=prefix)
-        imports.update({"import datetime", "from typing import cast", "from dateutil.parser import isoparse"})
-        return imports
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class DateProperty(Property):
-    """A property of type datetime.date"""
-
-    _type_string: ClassVar[str] = "datetime.date"
-    _json_type_string: ClassVar[str] = "str"
-    template: ClassVar[str] = "date_property.py.jinja"
-
-    def get_imports(self, *, prefix: str) -> Set[str]:
-        """
-        Get a set of import strings that should be included when this property is used somewhere
-
-        Args:
-            prefix: A prefix to put before any relative (local) module names. This should be the number of . to get
-            back to the root of the generated client.
-        """
-        imports = super().get_imports(prefix=prefix)
-        imports.update({"import datetime", "from typing import cast", "from dateutil.parser import isoparse"})
-        return imports
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class FileProperty(Property):
-    """A property used for uploading files"""
-
-    _type_string: ClassVar[str] = "File"
-    # Return type of File.to_tuple()
-    _json_type_string: ClassVar[str] = "FileJsonType"
-    template: ClassVar[str] = "file_property.py.jinja"
-
-    def get_imports(self, *, prefix: str) -> Set[str]:
-        """
-        Get a set of import strings that should be included when this property is used somewhere
-
-        Args:
-            prefix: A prefix to put before any relative (local) module names. This should be the number of . to get
-            back to the root of the generated client.
-        """
-        imports = super().get_imports(prefix=prefix)
-        imports.update({f"from {prefix}types import File, FileJsonType", "from io import BytesIO"})
-        return imports
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class FloatProperty(Property):
-    """A property of type float"""
-
-    _type_string: ClassVar[str] = "float"
-    _json_type_string: ClassVar[str] = "float"
-    _allowed_locations: ClassVar[Set[oai.ParameterLocation]] = {
-        oai.ParameterLocation.QUERY,
-        oai.ParameterLocation.PATH,
-        oai.ParameterLocation.COOKIE,
-        oai.ParameterLocation.HEADER,
-    }
-    template: ClassVar[str] = "float_property.py.jinja"
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class IntProperty(Property):
-    """A property of type int"""
-
-    _type_string: ClassVar[str] = "int"
-    _json_type_string: ClassVar[str] = "int"
-    _allowed_locations: ClassVar[Set[oai.ParameterLocation]] = {
-        oai.ParameterLocation.QUERY,
-        oai.ParameterLocation.PATH,
-        oai.ParameterLocation.COOKIE,
-        oai.ParameterLocation.HEADER,
-    }
-    template: ClassVar[str] = "int_property.py.jinja"
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class BooleanProperty(Property):
-    """Property for bool"""
-
-    _type_string: ClassVar[str] = "bool"
-    _json_type_string: ClassVar[str] = "bool"
-    _allowed_locations: ClassVar[Set[oai.ParameterLocation]] = {
-        oai.ParameterLocation.QUERY,
-        oai.ParameterLocation.PATH,
-        oai.ParameterLocation.COOKIE,
-        oai.ParameterLocation.HEADER,
-    }
-    template: ClassVar[str] = "boolean_property.py.jinja"
