+++ openapi_python_client/typing.py
@@ -1,7 +1,7 @@
 from typing import Callable, Set, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from parser.endpoints import EndpointCollection
+    from openapi_python_client.parser.endpoints import EndpointCollection
 
 
 TEndpointFilter = Callable[["EndpointCollection"], Set[str]]
